Add custom temperature sensor message
Not intended to replace flight controller messages (there isn't a
standard message that matches this use case anyway).

We will support up to 8 sensors, and users will have to label them
in QOpenHD in order to tell which ones are which.

When no sensor is detected for a particular slot, the value in the
message list will be left as -127C, which is not something a sensor
would be able to reach in real world use.

This will allow QOpenHD to show a "not  connected" status for that
sensor slot.